+++ vector/src/main/java/im/vector/riotx/features/autocomplete/group/AutocompleteGroupPresenter.kt
@@ -14,20 +14,20 @@
  * limitations under the License.
  */
 
-package im.vector.riotx.features.autocomplete.user
+package im.vector.riotx.features.autocomplete.group
 
 import android.content.Context
 import androidx.recyclerview.widget.RecyclerView
 import com.airbnb.mvrx.Async
 import com.airbnb.mvrx.Success
 import com.otaliastudios.autocomplete.RecyclerViewPresenter
-import im.vector.matrix.android.api.session.user.model.User
+import im.vector.matrix.android.api.session.group.model.GroupSummary
 import im.vector.riotx.features.autocomplete.AutocompleteClickListener
 import javax.inject.Inject
 
-class AutocompleteUserPresenter @Inject constructor(context: Context,
-                                                    private val controller: AutocompleteUserController
-) : RecyclerViewPresenter<User>(context), AutocompleteClickListener<User> {
+class AutocompleteGroupPresenter @Inject constructor(context: Context,
+                                                     private val controller: AutocompleteGroupController
+) : RecyclerViewPresenter<GroupSummary>(context), AutocompleteClickListener<GroupSummary> {
 
     var callback: Callback? = null
 
@@ -41,21 +41,21 @@ class AutocompleteUserPresenter @Inject constructor(context: Context,
         return controller.adapter
     }
 
-    override fun onItemClick(t: User) {
+    override fun onItemClick(t: GroupSummary) {
         dispatchClick(t)
     }
 
     override fun onQuery(query: CharSequence?) {
-        callback?.onQueryUsers(query)
+        callback?.onQueryGroups(query)
     }
 
-    fun render(users: Async<List<User>>) {
-        if (users is Success) {
-            controller.setData(users())
+    fun render(groups: Async<List<GroupSummary>>) {
+        if (groups is Success) {
+            controller.setData(groups())
         }
     }
 
     interface Callback {
-        fun onQueryUsers(query: CharSequence?)
+        fun onQueryGroups(query: CharSequence?)
     }
 }