+++ vector/src/main/java/im/vector/riotx/features/home/room/detail/composer/TextComposerAction.kt
@@ -20,4 +20,6 @@ import im.vector.riotx.core.platform.VectorViewModelAction
 
 sealed class TextComposerAction : VectorViewModelAction {
     data class QueryUsers(val query: CharSequence?) : TextComposerAction()
+    data class QueryRooms(val query: CharSequence?) : TextComposerAction()
+    data class QueryGroups(val query: CharSequence?) : TextComposerAction()
 }